Split off some UniqueType helper classes, make regions work in Android Studio and linting
