+++ core/src/com/unciv/models/ruleset/unique/UniqueFlag.kt
@@ -1,0 +1,9 @@
+package com.unciv.models.ruleset.unique
+
+enum class UniqueFlag {
+    HiddenToUsers,
+    ;
+    companion object {
+        val setOfHiddenToUsers = listOf(HiddenToUsers)
+    }
+}
